+++ app/src/main/java/com/example/playlistmaker/domain/favorites/FavoritesRepository.kt
@@ -7,5 +7,5 @@ interface FavoritesRepository {
     suspend fun addTrack(track: Track)
     suspend fun removeTrack(track: Track)
     fun getFavoriteTracks(): Flow<List<Track>>
-    suspend fun isFavorite(trackId: Int): Boolean
+    suspend fun isFavorite(trackId: Long): Boolean
 }
